refactor: rename and move listener functions
