+++ src/view/chat/commands/connection.ts
@@ -1,5 +1,5 @@
 import { Namespace, Socket } from 'socket.io'
-import { chatCommandListener, commandListener } from '../utils'
+import { commandListener, commandListenerWithVerification } from '../../utils'
 import { disconnectionCommand } from './disconnect'
 import { userTokenCommand } from './userToken'
 import { ChatController } from '../../../controllers/chat/chatController'
@@ -13,17 +13,13 @@ import { ChatController } from '../../../controllers/chat/chatController'
  * @param chatController
  */
 export function connectionCommand(chatNamespace: Namespace, chatController: ChatController) {
-  commandListener(
+  commandListenerWithVerification(
     chatNamespace,
     'connection',
     () => true,
     (socket: Socket) => {
-      chatCommandListener(
-        socket,
-        'userToken',
-        userTokenCommand(chatNamespace, socket, chatController)
-      )
-      chatCommandListener(socket, 'disconnect', disconnectionCommand(socket))
+      commandListener(socket, 'userToken', userTokenCommand(chatNamespace, socket, chatController))
+      commandListener(socket, 'disconnect', disconnectionCommand(socket))
     }
   )
 }
